My Groups: Get groups from user context rather than flux call
