+++ common/components/stores/UniversalDispatcher.js
@@ -2,7 +2,6 @@
 
 "use strict";
 
-import type { MyGroupsActionType } from "./MyGroupsStore.js";
 import type { NavigationActionType } from "./NavigationStore.js";
 import type { OffsetActionType } from "./PageOffsetStore.js";
 import type { LinkListActionType } from "./LinkListStore.js";
@@ -10,7 +9,6 @@ import type { FormFieldsActionType } from "./FormFieldsStore.js";
 import { Dispatcher } from "flux";
 
 type UniversalActionType =
-  | MyGroupsActionType
   | NavigationActionType
   | OffsetActionType
   | LinkListActionType
